+++ app/(client)/(anon)/spots/components/BottomSheet.module.scss
@@ -1,7 +1,117 @@
-.bottomSheetContainer {
+.bottomSheet {
+  position: fixed;
+  bottom: 0;
+  transform: translateX(-50%);
+  max-width: 30rem;
   width: 100%;
-  position: relative;
-  .hiddenSEO {
-    display: none; // SEO용 SSR 데이터 숨김
+  height: 12.5rem;
+  background: #fff;
+  border-radius: 1.25rem 1.25rem 0 0;
+  box-shadow: 0rem -0.25rem 0.625rem rgba(0, 0, 0, 0.1);
+  transition: transform 0.3s ease-in-out;
+  overflow: hidden;
+  z-index: 100;
+
+  &.open {
+    height: 72vh;
+    transform: translateX(-50%) translateY(0);
+  }
+
+  .wrapHandle {
+    display: flex;
+    align-items: center;
+    width: 100%;
+    height: 2.5rem;
+    cursor: pointer;
+    .handle {
+      width: 3.125rem;
+      height: 0.3125rem;
+      margin: 0 auto;
+      background: var(--disabled-color);
+      border-radius: 0.3125rem;
+    }
+  }
+  .noSpots {
+    text-align: center;
+    padding: 20px;
+    font-size: 16px;
+    font-weight: bold;
+    color: #777;
+  }
+
+  .list {
+    padding: 0 1.25rem;
+    overflow-y: auto;
+    max-height: calc(69vh - 2.5rem);
+
+    .spotItem {
+      width: 100%;
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
+      padding: 0.625rem;
+      border-bottom: 0.0625rem solid var(--grey-2-color);
+
+      .info {
+        flex: 1;
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        margin-right: 0.625rem;
+
+        .top {
+          display: flex;
+          .name {
+            text-decoration: none;
+            color: black;
+            font-size: 1.25rem;
+            font-weight: bold;
+            margin-right: 0.625rem;
+          }
+
+          .category {
+            line-height: 1.25rem;
+            color: var(--grey-2-color);
+          }
+        }
+
+        .distance-time {
+          display: flex;
+          margin: 0.9375rem 0;
+          .distance {
+            color: var(--primary-color);
+            margin-right: 0.625rem;
+          }
+        }
+
+        .details {
+          width: 90%;
+          display: flex;
+          justify-content: space-between;
+          align-items: center;
+          font-size: 1rem;
+
+          .icon {
+            font-size: 1.5625rem;
+            color: var(--primary-color);
+          }
+          .bookmark,
+          .visitors,
+          .price {
+            display: flex;
+            align-items: center;
+            gap: 0.3125rem;
+            color: #444;
+          }
+        }
+      }
+      .image {
+        width: 6.25rem;
+        height: 6.25rem;
+        border-radius: 0.625rem;
+        object-fit: cover;
+        flex-shrink: 0;
+      }
+    }
   }
 }
